Fix RestTemplate support
