+++ apm-agent-plugins/apm-spring-resttemplate/apm-spring-resttemplate-plugin/src/test/java/co/elastic/apm/agent/resttemplate/SpringRestTemplateInstrumentationTest.java
@@ -103,10 +103,6 @@ public class SpringRestTemplateInstrumentationTest extends AbstractHttpClientIns
                 // We do not support body capturing for OkHttp yet
                 return false;
             }
-            if (restTemplate.getRequestFactory() instanceof HttpComponentsClientHttpRequestFactory) {
-                //apache http client v5 is also not supported yet
-                return false;
-            }
             return true;
         }
